pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harness-lab"
version = "1.0.0"
description = "研究优先的 Harness Lab 平台"
authors = [
    {name = "Harness Lab Team"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.23",
    "psycopg[binary]>=3.1.18",
    "redis>=5.0.1",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.3",
    "sentence-transformers>=2.2.2",
    "openai>=1.3.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "transformers>=4.35.0",
    "torch>=2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=5.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/your-username/harness-lab"
Repository = "https://github.com/your-username/harness-lab.git"
Documentation = "https://github.com/your-username/harness-lab#readme"

[project.scripts]
harness-lab = "backend.app.main:main"
hlab = "backend.app.harness_lab.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
